refactor: move nested to nested_impl.package and create nested lib
The new "nested" lib shall contain all utility functions related
to manipulating nested data structures.
